+++ packages/gatsby-plugin-feed/src/gatsby-node.js
@@ -100,7 +100,7 @@ exports.onPostBuild = async ({ graphql }, pluginOptions) => {
         ? feed.serialize
         : serialize
 
-    const rssFeed = serializer(locals).reduce((merged, item) => {
+    const rssFeed = (await serializer(locals)).reduce((merged, item) => {
       merged.item(item)
       return merged
     }, new RSS(setup(locals)))